Fix partially hidden words issue
If one word partially overlaps with another word (e.g. "message"
partially contains "gear"), then there's a chance that it can be hard
to later detect where the first word ends and the second word starts.
This commit limits the allowed number of consequentially overlapping
characters to just one character, which should make it easier to
distinguish the word boundaries.